+++ src/components/TableHeader/TableHeader.tsx
@@ -1,10 +1,10 @@
 import { useCallback, useMemo } from 'react'
 import { OrderBy, toggleColumn } from '../../helpers/sort.js'
-import { ColumnDescriptor } from '../../hooks/useTableConfig.js'
+import { ColumnParameters } from '../../hooks/useTableConfig.js'
 import ColumnHeader from '../ColumnHeader/ColumnHeader.js'
 
 interface TableHeaderProps {
-  columnDescriptors: ColumnDescriptor[]
+  columnsParameters: ColumnParameters[]
   orderBy?: OrderBy // array of column order by clauses. If undefined, the table is unordered, the sort elements are hidden and the interactions are disabled.
   onOrderByChange?: (orderBy: OrderBy) => void // callback to call when a user interaction changes the order. The interactions are disabled if undefined.
   canMeasureWidth: boolean // indicates if the width of the columns can be measured.
@@ -16,7 +16,7 @@ interface TableHeaderProps {
  * Render a header for a table.
  */
 export default function TableHeader({
-  columnDescriptors, orderBy, onOrderByChange, canMeasureWidth, ariaRowIndex, columnClassNames = [],
+  columnsParameters, orderBy, onOrderByChange, canMeasureWidth, ariaRowIndex, columnClassNames = [],
 }: TableHeaderProps) {
   // Function to handle click for changing orderBy
   const getToggleOrderBy = useCallback((columnHeader: string) => {
@@ -30,8 +30,8 @@ export default function TableHeader({
     return new Map((orderBy ?? []).map(({ column, direction }, index) => [column, { direction, index }]))
   }, [orderBy])
 
-  return columnDescriptors.map((colDescriptor) => {
-    const { key: name, index: columnIndex, ...columnConfig } = colDescriptor
+  return columnsParameters.map((columnParameters) => {
+    const { name, index: columnIndex, ...columnConfig } = columnParameters
     // Note: columnIndex is the index of the column in the dataframe header
     // and not the index of the column in the table (which can be different if
     // some columns are hidden, or if the order is changed)