+++ src/helpers/dataframe/array.ts
@@ -1,9 +1,10 @@
 import { validateColumn, validateRow } from './helpers.js'
-import type { ResolvedValue, UnsortableDataFrame } from './types.js'
+import type { DataFrame, ResolvedValue } from './types.js'
+import { type OrderBy, validateOrderBy } from '../sort.js'
 
-export function arrayDataFrame(array: Record<string, any>[], rowNumbers?: number[]): UnsortableDataFrame {
+export function arrayDataFrame(array: Record<string, any>[], rowNumbers?: number[]): DataFrame {
   // beware: we don't copy the array, so it must not be mutated after this point.
-  const header = 0 in array ? Object.keys(array[0]) : []
+  const columnDescriptors = (0 in array ? Object.keys(array[0]) : []).map(name => ({ name }))
   const numRows = array.length
   if (rowNumbers && rowNumbers.length !== numRows) {
     throw new Error(`Row numbers length (${rowNumbers.length}) does not match the number of rows in the array (${numRows})`)
@@ -13,8 +14,9 @@ export function arrayDataFrame(array: Record<string, any>[], rowNumbers?: number
     throw new Error(`Row numbers must be non-negative integers, but got: ${rowNumbers.join(', ')}`)
   }
 
-  function getRowNumber({ row }: { row: number }): ResolvedValue<number> {
+  function getRowNumber({ row, orderBy }: { row: number, orderBy?: OrderBy }): ResolvedValue<number> {
     validateRow({ row, data: { numRows } })
+    validateOrderBy({ orderBy })
     if (!rowNumbers) {
       return { value: row }
     }
@@ -24,9 +26,10 @@ export function arrayDataFrame(array: Record<string, any>[], rowNumbers?: number
     return { value: rowNumbers[row] }
   }
 
-  function getCell({ row, column }: { row: number, column: string }): ResolvedValue | undefined {
+  function getCell({ row, column, orderBy }: { row: number, column: string, orderBy?: OrderBy }): ResolvedValue | undefined {
     validateRow({ row, data: { numRows } })
-    validateColumn({ column, data: { header } })
+    validateColumn({ column, data: { columnDescriptors } })
+    validateOrderBy({ orderBy })
     const cells = array[row]
     if (!cells) {
       throw new Error(`Row ${row} not found in data`)
@@ -39,5 +42,5 @@ export function arrayDataFrame(array: Record<string, any>[], rowNumbers?: number
     return { value: cells[column] }
   }
 
-  return { numRows, header, getRowNumber, getCell }
+  return { numRows, columnDescriptors, getRowNumber, getCell }
 }